# backend/openapi/schema.yaml
openapi: 3.0.0
basePath: '/v1'
info:
  title: 'Sample API'
  description: 'sample'
  version: 0.1.0
components:
  schemas:
    File:
      type: object
      properties:
        name:
          description: File identifier
          type: string
        root:
          description: Root directory of file, denotes intended use.
          type: string
        file_path:
          description: Local path inside docker file mapping.
          type: string
        url_path:
          description: Remote path for file, fetched at event export.
          type: string
...
    Theme:
      type: object
      properties:
        name:
          description: Theme identifier.
          type: string
        overlay_file:
          description: Identifier for overlay theme file.
          type: string
        stinger_file:
          description: Identifier for stinger theme file.
          type: string
        starting_file:
          description: Identifier for starting theme file.
          type: string
        ending_file:
          description: Identifier for ending theme file.
          type: string
        target_video_width:
          description: Pixel width for video files to be scaled to.
          type: integer
          format: int32
        target_video_height:
          description: Pixel height for video files to be scaled to.
          type: integer
          format: int32
        video_offset_x:
          description: Pixel count for video files to horizontally offset by.
          type: integer
          format: int32
        video_offset_y:
          description: Pixel count for video files to vertically offset by.
          type: integer
          format: int32
        chat_width:
          description: Pixel width for overlay chat.
          type: integer
          format: int32
        chat_height:
          description: Pixel height for overlay chat.
          type: integer
          format: int32
        chat_offset_x:
          description: Pixel count for overlay chat to horizontally offset by.
          type: integer
          format: int32
        chat_offset_y:
          description: Pixel count for overlay chat to vertically offset by.
          type: integer
          format: int32
...
    DJ:
      type: object
      properties:
        name:
          description: DJ identifier. Used in streams if no logo or public_name is provided.
          type: string
        logo:
          description: Logo file identifier.
          type: string
        recording:
          description: Recording file identifier.
          type: string
        rtmp_server:
          description: Specific RTMP server to use for live streams.
          type: string
        rtmp_key:
          description: RTMP key to use for live streams.
          type: string
        public_name:
          description: Public name for displaying in stream. Used if no logo is provided.
          type: string
        discord_id:
          description: Discord ID, used for bookkeeping and contact.
          type: string
        past_events:
          description: List of past events the DJ has performed in
          type: array
          items:
            type: string
...
    LineupDJ:
      type: object
      properties:
        name:
          description: DJ identifier
          type: string
        is_live:
          description: If the DJ is using a RTMP config for streaming.
          type: boolean
        vj:
          description: Present if the visuals for a DJ are done by a VJ.
          type: string
    UpdateLineupDJ:
      type: object
      properties:
        is_live:
          description: If the DJ is using a RTMP config for streaming.
          type: boolean
        vj:
          description: Present if the visuals for a DJ are done by a VJ.
          type: string
    LineupPromotion:
      type: object
      properties:
        name:
          description: Name of an exsting Promotion.
          type: string
    Event:
      type: object
      properties:
        name:
          description: Event identifier
          type: string
        djs:
          description: Array of LineupDJ objects for an event, describes event specific values.
          type: array
          items:
            $ref: '#/components/schemas/LineupDJ'
        promos:
          description: Array of promotional video identifiers.
          type: array
          items:
            type: string
        theme:
          description: Theme identifier to use for an event.
          type: string
        public:
          description: If the event has been announced.
          type: boolean
        date:
          description: Date the event is to be aired, in YYYY/MM/DD.
          type: string
        start_time:
          description: Time the event is to be aired, in 24hr HH:MM for eastern time.
          type: string
    UpdateEvent:
      type: object
      properties:
        djs:
          description: Array of LineupDJ objects for an event, describes event specific values.
          type: array
          items:
            $ref: '#/components/schemas/LineupDJ'
        promos:
          description: Array of promotional video identifiers.
          type: array
          items:
            type: string
        theme:
          description: Theme identifier to use for an event.
          type: string
        public:
          description: If the event has been announced.
          type: boolean
        date:
          description: Date the event is to be aired, in YYYY/MM/DD.
          type: string
        start_time:
          description: Time the event is to be aired, in 24hr HH:MM for eastern time.
          type: string
    UpdateEventDateTime:
      type: object
      properties:
        date:
          description: Date the event is to be aired, in YYYY/MM/DD.
          type: string
        start_time:
          description: Time the event is to be aired, in 24hr HH:MM for eastern time.
          type: string
    MoveEventItem:
      type: object
      properties:
        index_a:
          description: Integer of the item to move.
          type: integer
          format: int32
        index_b:
          description: Integer to move the item to.
          type: integer
          format: int32
    EventThemeParam:
      type: object
      properties:
        name:
          description: Name of an exsting theme.
          type: string
    Events:
      type: array
      items:
        $ref: '#/components/schemas/Event'
...
    EventExportSummary:
      type: object
      properties:
        event:
          description: Event object.
          $ref: '#/components/schemas/Event'
        djs:
          description: Array of full DJ objects for an event.
          type: array
          items:
            $ref: '#/components/schemas/DJ'
        promos:
          description: Array of promotional video objects.
          type: array
          items:
            $ref: '#/components/schemas/Promo'
        theme:
          description: Theme object for an event.
          $ref: '#/components/schemas/Theme'
        files:
          description: Array of file objects for DJ and Promotion media.
          type: array
          items:
            $ref: '#/components/schemas/File'
    Promo:
      type: object
      properties:
        name:
          description: Promo identifier
          type: string
        promo_file:
          description: File identifier.
          type: string
...
paths:
...
  /event:
    get:
      tags:
        - events
      summary: Returns a list of all events.
      responses:
        '200':
          description: An array of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Events'
    post:
      tags:
        - events
      summary: Create a new event.
      requestBody:
        description: New event values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
        required: true
      responses:
        '200':
          description: New event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid input data.
        '409':
          description: Event already exists.
...
  /event/{eventName}:
    get:
      tags:
        - events
      summary: Returns a event by name
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
    post:
      tags:
        - events
      summary: Update existing event.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Updated event values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvent'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
    delete:
      tags:
        - events
      summary: Delete existing event.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Deleted event.
        '404':
          description: Event not found.
  /event/{eventName}/dateTime:
    post:
      tags:
        - events
      summary: Update an event's date and time.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Updated date and time values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventDateTime'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/dj:
    post:
      tags:
        - events
      summary: Add a DJ to an event's lineup.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: DJ lineup values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineupDJ'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/dj/{djName}:
    post:
      tags:
        - events
      summary: Update an event's lineup DJ.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
        - name: djName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Updates to the lineup DJ.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLineupDJ'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
    delete:
      tags:
        - events
      summary: Removes a DJ from an event's lineup.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
        - name: djName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/move-dj:
    post:
      tags:
        - events
      summary: Moves a DJ in an event lineup.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Indexes to move.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveEventItem'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/promo:
    post:
      tags:
        - events
      summary: Add a promotion to an event.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Promotion values.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineupPromotion'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/promo/{promoName}:
    delete:
      tags:
        - events
      summary: Removes a Promotion from an event's lineup.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
        - name: promoName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Removed Promotion from event.
        '404':
          description: Event not found.
  /event/{eventName}/move-promo:
    post:
      tags:
        - events
      summary: Moves a Promotion in an event lineup.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Indexes to move.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveEventItem'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/set-theme:
    post:
      tags:
        - events
      summary: Sets the theme for an event.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      requestBody:
        description: Theme name.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventThemeParam'
        required: true
      responses:
        '200':
          description: Updated event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
  /event/{eventName}/export:
    post:
      tags:
        - events
      summary: Exports an event to a local json file.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: Exported event.
        '404':
          description: Event not found.
        '409':
          description: Event export failed.
  /event/{eventName}/export-summary:
    get:
      tags:
        - events
      summary: Prepares a summary of an event's export.
      parameters:
        - name: eventName
          in: path
          required: true
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: All objects related to exporting the event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventExportSummary'
        '404':
          description: Event not found.
...